fix(census): use real census size from bundle for CSP elections
For CSP (OFF_CHAIN_CA) elections the SDK builds a CspCensus without a
size, so census size and participation fell back to maxCensusSize, which
is the allowed-voters cap rather than the actual census size. The real
size lives in the bundle JSON at census.censusURI.

Add useCensusSize() (and promote useCensusBundle to a shared query) that
reads the size from the bundle for CSP elections and falls back to
census.size/maxCensusSize otherwise. Wire it into Summary, View and the
dashboard ProcessView, recomputing participation locally where the
react-components hook value would otherwise be wrong.
