+++ src/components/Process/CSP/CSPAuthModal.tsx
@@ -2,7 +2,7 @@ import { Button, CloseButton, Dialog, Portal } from '@chakra-ui/react'
 import { useElection } from '@vocdoni/react-components'
 import { InvalidElection, PublishedElection } from '@vocdoni/sdk'
 import { Trans, useTranslation } from 'react-i18next'
-import { useCensusBundle } from '~queries/bundle'
+import { useCensusBundle } from '~queries/census'
 import { CspAuthProvider, useCspAuthContext } from './CSPStepsProvider'
 import { Step0Base } from './Step0'
 import { Step1Base } from './Step1'
