+++ src/components/Process/View.tsx
@@ -30,12 +30,15 @@ import {
   useElection,
   useOrganization,
 } from '@vocdoni/react-components'
-import { CensusType, ElectionStatus, PublishedElection } from '@vocdoni/sdk'
+import { ElectionStatus, PublishedElection } from '@vocdoni/sdk'
 import { ReactNode, useEffect, useRef, useState } from 'react'
 import { Trans, useTranslation } from 'react-i18next'
 import { RiBarChartBoxLine, RiErrorWarningLine } from 'react-icons/ri'
 import { generatePath } from 'react-router-dom'
 import { Routes } from '~src/router/routes'
+import { RiErrorWarningLine } from 'react-icons/ri'
+import { useCensusSize } from '~queries/census'
+import { FacebookShare, RedditShare, TelegramShare, TwitterShare } from '~components/Share'
 import { BallotBoxAnimated } from '../Layout/BallotBoxAnimated'
 import { ActionsMenu } from './ActionsMenu'
 import ProcessAside, { VoteButton } from './Aside'
@@ -45,8 +48,6 @@ import { ElectionVideo } from './Dashboard/ProcessView'
 import { ProcessDate } from './Date'
 import Header from './Header'
 import { useVotingMethodLabel } from './resultTypeLabels'
-
-type CensusInfo = { size: number; weight: bigint; type: CensusType }
 
 type ProcessInfoCardProps = {
   label: string
@@ -89,27 +90,15 @@ const ProcessInfoPanel = () => {
   const { t } = useTranslation()
   const { election } = useElection()
   const { organization, loaded } = useOrganization()
-  const { account, client } = useClient()
-  const [censusInfo, setCensusInfo] = useState<CensusInfo>()
-
-  // Get the census info to show the total size if the maxCensusSize is less than the total size
-  useEffect(() => {
-    ;(async () => {
-      try {
-        if (!client || !(election instanceof PublishedElection) || !election?.census?.censusId) return
-        const censusInfo: CensusInfo = await client.fetchCensusInfo(election.census.censusId)
-        setCensusInfo(censusInfo)
-      } catch (e) {
-        // If the census info is not available, just ignore it
-        setCensusInfo(undefined)
-      }
-    })()
-  }, [election, client])
+  const { account } = useClient()
+  // For CSP elections the actual census size comes from the bundle, not from `maxCensusSize`
+  // (which only caps how many voters may vote). See useCensusSize.
+  const { size: censusSize } = useCensusSize()
 
   if (!(election instanceof PublishedElection)) return null
 
   const showOrgInformation = !loaded || (loaded && organization?.account?.name)
-  const showTotalCensusSize = censusInfo?.size && election?.maxCensusSize && election.maxCensusSize < censusInfo.size
+  const showTotalCensusSize = censusSize > 0 && !!election?.maxCensusSize && election.maxCensusSize < censusSize
 
   return (
     <Flex
@@ -143,7 +132,7 @@ const ProcessInfoPanel = () => {
               <TooltipTrigger asChild>
                 <Text>
                   {t('process.total_census_size', {
-                    censusSize: censusInfo?.size,
+                    censusSize,
                     maxCensusSize: election?.maxCensusSize,
                   })}
                 </Text>
@@ -151,11 +140,11 @@ const ProcessInfoPanel = () => {
               <TooltipPositioner>
                 <TooltipContent bg='primary.600' color='white'>
                   {t('process.total_census_size_tooltip', {
-                    censusSize: censusInfo?.size,
+                    censusSize,
                     maxCensusSize: election?.maxCensusSize,
                     percent:
-                      censusInfo?.size && election?.maxCensusSize
-                        ? Math.round((election?.maxCensusSize / censusInfo?.size) * 100)
+                      censusSize && election?.maxCensusSize
+                        ? Math.round((election?.maxCensusSize / censusSize) * 100)
                         : 0,
                   })}
                 </TooltipContent>
@@ -163,7 +152,7 @@ const ProcessInfoPanel = () => {
             </TooltipRoot>
           ) : (
             <Text color='texts.subtle' fontSize='sm'>
-              {t('process.people_in_census', { count: election?.maxCensusSize })}
+              {t('process.people_in_census', { count: censusSize })}
             </Text>
           )
         }
